Se tienen en cuentalos tipos de datos de mongoId http://mongoid.org/en/mongoid/docs/documents.html#fields

+++ app/models/user_history.rb
@@ -6,5 +6,5 @@ class UserHistory
   field :dependence, type: String
   field :priotity, type: Integer
   field :iteration, type: String
-  field :true_time, type: Decimal
+  field :true_time, type: Float
 end
